+++ tests/api/txs_test.go
@@ -1,4 +1,4 @@
-package tests
+package api
 
 import (
 	"testing"
